requirements.txt
asgiref==3.3.1
audio2numpy==0.1.2
boto3==1.17.33
botocore==1.20.33
cffi==1.14.5
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.7
django-cleanup==5.1.0
django-heroku==0.3.1
django-storages==1.11.1
djangorestframework==3.12.2
ffmpeg==1.4
gunicorn==20.0.4
jmespath==0.10.0
kiwisolver==1.3.1
matplotlib==3.2.0
numpy==1.20.1
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.6
six==1.15.0
SoundFile==0.10.3.post1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
